--- a/schema.sql
+++ b/schema.sql
@@ -48,3 +48,4 @@ CREATE TABLE visits(
 
 CREATE INDEX idx_animal_id ON visits (animal_id);
 CREATE INDEX idx_visits_vet_id ON visits (vet_id);
+CREATE INDEX idx_owners_email ON owners (email);