--- a/schema.sql
+++ b/schema.sql
@@ -5,6 +5,9 @@ CREATE TABLE owners(
   PRIMARY KEY(id)
 );
 
+-- Add an email column to your owners table
+ALTER TABLE owners ADD COLUMN email VARCHAR(120);
+
 CREATE TABLE species(
   id INT GENERATED ALWAYS AS IDENTITY,
   name VARCHAR(20),