… screenshots/query-1-after.png
… screenshots/query-1-before.png
… screenshots/query-2-after.png
… screenshots/query-2-before.png
… screenshots/query-3-after.png
… screenshots/query-3-before.png