### Overview
This pull request aims to optimize the performance of specific queries in the Vet Clinic Database, as outlined in the project requirements. The focus was reducing the execution time for queries that took significantly longer to execute (more than 1000ms).

**Note: Below are the screenshots attached for the reason I'm working solo.** 
![image](https://github.com/GhausAliKhan/vet-clinic/assets/94997336/9b87613e-6762-4a4d-b69c-f7be003297b8)
![image](https://github.com/GhausAliKhan/vet-clinic/assets/94997336/1af8950b-1143-4975-bae6-efa8be3e6a3c)

### Changes Made

- Add an index on the `animal_id` column in the `visits` table to optimize `SELECT COUNT(*) FROM visits where animal_id = 4;`.
- Add an index on the `vet_id` column in the `visits` table to optimize `SELECT * FROM visits where vet_id = 2;`.
- Add an index on the `email` column in the `owners` table to optimize `SELECT * FROM owners where email = '[owner_18327@mail.com](mailto:owner_18327@mail.com)';`.

### Performance Improvements
The `EXPLAIN ANALYZE` screenshots below show the execution time before and after the changes:

1. Query: `SELECT COUNT(*) FROM visits where animal_id = 4;`

     - Before Execution time was 346.540 ms.
![query-1-before](https://github.com/GhausAliKhan/vet-clinic/assets/94997336/6c8b882a-17eb-4fc4-849f-c51f4413fb8c)
     - After Execution time was reduced to 122.850 ms.
![query-1-after](https://github.com/GhausAliKhan/vet-clinic/assets/94997336/5799e264-cd2b-450d-9482-6ffc344fd37f)

2. Query: `SELECT * FROM visits where vet_id = 2;`

     - Before Execution time was 494.496 ms.
![query-2-before](https://github.com/GhausAliKhan/vet-clinic/assets/94997336/2c6d37e7-d761-408a-89cc-0a032c38bdd0)
     - After Execution time was 411.002 ms.
![query-2-after](https://github.com/GhausAliKhan/vet-clinic/assets/94997336/898ab013-73ba-4b9e-b068-6474d3db655c)

3. Query: `SELECT * FROM owners where email = '[owner_18327@mail.com](mailto:owner_18327@mail.com)';`

     - Before Execution time was 341.304 ms.
![query-3-before](https://github.com/GhausAliKhan/vet-clinic/assets/94997336/5c834674-1213-4110-8328-2b0a835b4da8)
     - After Execution time was 0.082 ms. 
![query-3-after](https://github.com/GhausAliKhan/vet-clinic/assets/94997336/0a5177c3-5447-4c34-8c81-90825384a3c9)

### SQL Files

- Updated schema.sql to include the newly added indexes.
- No changes were made to the data.sql file as the structure of the database remains the same.